requirements.txt
Flask
Flask_SocketIO==5.1.0
gevent==21.1.2
gevent-websocket==0.10.1
gunicorn
numpy
spotipy
python-dotenv